Graphs from Samuel Norman Seaborn
The Nobel Prize has been among the most prestigious international awards since 1901. Awards are bestowed in chemistry, literature, physics, physiology or medicine, economics, and peace each year. In addition to the honor, prestige, and substantial prize money, the recipient also gets a gold medal with an image of Alfred Nobel (1833 - 1896), who established the prize.

The Nobel Foundation has made a dataset available of all prize winners from the outset of the awards from 1901 to 2023. The dataset used in this project is from the Nobel Prize API and is available in the `nobel.csv` file in the `data` folder.

In this project, you'll explore and answer several questions related to this prizewinning data.